fix: error out the stx protocol if the sender sends unsupported data (#5572)
Description
---
Will error out the stx protocol if the sender sends unsupported script,
features, or covenant

Motivation and Context
---
Its possible for the sender to block the spending of the transaction by
using unsupported script, covenant or features. This will now make it
that the receiver will not accept a transaction it does not know how to
spend.

How Has This Been Tested?
---
Passes all unit tests.