fix: minor fixes for multiple address support (#5617)
Description
---
Fixed a few places in the code where multiple addresses were not
completely supported, namely
- wallet base node detection via GRPC, and 
- whoami QR code

Fixed incorrect format for QR Code

Replaces replace_public_address with set_public_addresses.

Motivation and Context
---
Minor fixes to use all addresses rather than a single address.

With set_public_addresses, I wanted to set the addresses to only the
public addresses. However, this is problematic because the tor address
is added at a very late point in the comms initialization. This would
mean that if public addresses were configured along with a
late-configured tor address, the identity sig would change every time.

Other than in the comms builder, the tor hidden service code is
completely decoupled from comms (that is, we could pass in the transport
to comms and not the hidden service abstraction itself). In order to
solve this we would setup the tor HS earlier before we initialize the
node identity, and remove the code that mutates the node identity in the
comms builder. We'd know the final tor address and would be able to add
it to the public address list and have a final address set for node
identity. Although I anticipate this refactor would be fairly
straightforward, is not strictly necessary at this point so is not
attempted.

How Has This Been Tested?
---
Cucumbers

What process can a PR reviewer use to test or verify this change?
---
Add multiple addresses to the base node and check that the wallet is
aware of those addresses.

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>