feat: change default script to PushPubKey (#5653)
Description
---
Change the default script to `PushPubKey` to ensure the spending of
outputs from standard transactions is only performed with the use of
specific private keys.

Motivation and Context
---
This change to the default helps enforce that UTXOS can't be spent from
a software wallet, when a hardware wallet is in use.

How Has This Been Tested?
---
Manually, locally, but pretty tests still require updating.

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>