fix: memory overflow panic (#5658)
Description
---
Limits the varint size that the merkle_tree, covenent, script and stack
can be.

Motivation and Context
---
`fn deserialize_reader<R>(reader: &mut R) -> Result<Self, io::Error>`
trusts that the len of the varint declaring the size of the vec<u8> it
needs to read is honest. This can cause a node to comment up to u64::MAX
memory and cause a memory overflow. This adds a hard limit to what the
size can be and checks it before allocating the memory.


How Has This Been Tested?
---
Unit tests

Audit Finding Number
--
TARI-020

---------

Co-authored-by: stringhandler <stringhandler@gmail.com>