fix(comms)!: use noise XX handshake pattern for improved privacy (#5696)
Description
---
Replaces the noise IX handshake with the XX handshake
Adds a prologue to the noise hash state
Update snow from 0.9.1 to 0.9.3 (correctness issue fixed:
https://github.com/mcginty/snow/releases/tag/v0.9.2)

Motivation and Context
---

http://www.noiseprotocol.org/noise.html

The IX protocol provides no privacy for the initiator static key. We
replace it with XX which (essentially) encrypts the initiator static key
with an ephemeral Diffie-Hellman before sending the initiator static
public key at the cost of an extra send/recv i.e 1.5 RTT for XX vs 1 RTT
for IX. This provides forward secrecy for the initiator but weaker
privacy for the responder (Encrypted with forward secrecy, but can be
probed by an anonymous initiator.). This shouldn't be an issue as any
initiator can perform a valid connection and get the peer information
anyway,

An alternative is XK (initiator key is transmitted, responder key is
known) which does fit our requirements and has stronger privacy and
similar security guarantees, however is a 2 RTT pattern

A potential way to offset the increased RTT by -0.5 is to include
identity information usually performed as a separate protocol directly
after noise, as part of the noise handshake.

How Has This Been Tested?
---
Existing comms tests and all tests using comms

What process can a PR reviewer use to test or verify this change?
---
Create a localnet and check that nodes connect

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Cannot connect to peers without this change

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->