fix: better timeout for lagging (#5705)
Description
---
Removes some very specific off by one check, and instead use a
configured timeout before considering lagging.

This should also prevent a bug where a node knows of a stronger chain,
but it is at a lower height and this would cause it to think it was up
to date

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->