fix(comms): greatly reduce timeouts for first byte and noise handshake (#5728)
Description
---
Reduce time to first byte timeout from 45s to 3s
Remove noise protocol timeout
Add timeout for receiving noise handshake replies

Motivation and Context
---
Time to first byte should be short to prevent a malicious connection
from holding many open for long periods.

The "global" noise handshake timeout was excessively long (40s), this is
replaced with a timeout on the receive calls inside
the NoiseSocket. This means that the responder in the XX handshake will
wait for up to 6s (2 x receives) while the initiator will wait for 3s (1
receive). This should be more robust and reflect that the wait time for
a responder is generally larger than for the initiator.

How Has This Been Tested?
---
Existing tests, manually

What process can a PR reviewer use to test or verify this change?
---
Connections work!
<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->