fix!: remove timestamp from header in proto files (#5667)
Description
---
The protobuf definition of BlockHeader was using a `timestamp` type
which included `nanoseconds`. This data was discarded anyway and only
the seconds were included. I've deleted these and made the conversions
simpler

Motivation and Context
---
Removing extra data from a packet 

How Has This Been Tested?
---
Existing tests

What process can a PR reviewer use to test or verify this change?
---
Since this data was discarded, it was only really up to a malicious
party to populate this data. This has now been removed.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [x] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->
BREAKING CHANGE: Previous BlockHeader protobuf messages are incompatible