fix: prevent access violation when running multiple vms at the same time (#5734)
Description
---
Add a rwlock around the randomx virtual machine

Motivation and Context
---
Calling calculate hash on a vm while it is still busy with another
calculate hash would result in a panic. The unit test I created
recreated that problem.

How Has This Been Tested?
---
New test to reproduce the error

What process can a PR reviewer use to test or verify this change?
---
Apply the unit test to the development branch without the changes and
the test should fail

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->