fix: stuck on sync (#5739)
Description
---
Fixes the block sync and horizon sync to stop a node from getting stuck
on sync

Motivation and Context
---
When a local node has an error in talking to a node on sync, it should
not try over and over again, it should remove that peer from the list of
peers and try another peer. If that fails it should go back to listing
mode and get a new peer list.

Currently the node will get stuck and keep trying the same thing over
again without any changes being made locally to fix the problem.

```
07:57 WARN  RPC request failed: NotFound: Requested end block sync hash was not found
07:57 WARN  This sync round failed (450)
```
In this example, the node kept asking the same peer for the same block
hash it did not have. And thus it was stuck in sync mode until manual
intervention.

How Has This Been Tested?
---
manual

---------

Co-authored-by: Hansie Odendaal <39146854+hansieodendaal@users.noreply.github.com>