fix(comms+dht): mark peers as online inbound connection,join (#5741)
Description
---
- mark peers as online inbound connection and on join msg
- fix DHT peer query logic

Motivation and Context
---
Peers were not being marked as online whenever we have proof that they
are online, specifically in the inbound connection and join msg.

The peer query to exclude peers if we have attempted them but never
succeeded was incorrect and filtered out most peers. This was changed to

```rust
 // we have tried to connect to this peer, and we have never made a successful attempt at connection
if peer.last_connect_attempt().is_some() && peer.last_seen().is_none() {
    return false;
}
```

How Has This Been Tested?
---
Manually, node reconnects quickly to peers on startup

What process can a PR reviewer use to test or verify this change?
---

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->