feat(comms): allow multiple messaging protocol instances (#5748)
Description
---
Allow multiple message protocol instances

Motivation and Context
---
In L2 we want to have 2 messaging substreams, one for hotstuff messages
and one for the rest (transactions, network messages) to prevent the
queue of incoming transactions from causing HS leader failures.

In the base layer we spawn a single messaging protocol with the same
identitfier, so nothing changes
 
How Has This Been Tested?
---
Existing tests and manually

What process can a PR reviewer use to test or verify this change?
---
Node and wallet function as before

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

Co-authored-by: SW van Heerden <swvheerden@gmail.com>