feat: cli add list of vns for next epoch (#5743)
Description
---
Add list of VNs for next epoch.

Motivation and Context
---

How Has This Been Tested?
---
I used dan-testing to create VNs and to check if they are printed for
the next epoch as well.

What process can a PR reviewer use to test or verify this change?
---
Same as above.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->