Make the metadata update less evil
As in don't swap pointers out from under anyone. I was having a lot of
trouble validating the metadata was being applied to an existing message
via mutation but have since solved these issues. Now that we've verified
the mutation we no longer need to destroy and create new Message
objects.