fix: change truncate_from_bits to from_bits (#5773)
Description
---
This removes all occurrences of `truncate_from_bits` to `from_bits`

Motivation and Context
---
`truncate_from_bits` will truncate all unknown bits and may cause bits
to be interpreted as the wrong flag. This changes to the much more
strict `from_bits` which forces peers to use the correct bits and only
the correct bits.