chore: move mempool fee check (#5777)
Description
---
Moves the mempool fee check to before validation

Motivation and Context
---
This is a very cheap check compared to full validation. We should do
this before we check if the tx is valid so that we don't do validation
and then through away the tx because the fee is too low.