fix: possible exception in request_context (#5784)
Description
---
Fixed possible exception in `pub struct RequestContext` that could occur
when splitting the request and reply channel in `pub fn split`

Motivation and Context
---
The code should not be allowed to panic in use in the wrong context.

How Has This Been Tested?
---
Unit tests

What process can a PR reviewer use to test or verify this change?
---
Code walk through

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->

---------

Co-authored-by: SW van Heerden <swvheerden@gmail.com>