fix(chatffi): return and read from ptrs (#5827)
Description
---
A unit struct was being returned within a c_rep struct. This caused the
unit struct to have no body in the header file and caused build errors
in ios. This unit struct should have been a pointer with read functions
and not returned as a rust unit struct. This PR moves it over to a
pointer and corrects it's readers.

Motivation and Context
---
Don't produce broken builds. 

A lot of types _around_ this fix will be removed soon after this is
merged but it this corrects a broken build and it would be nicer to work
on top of this PR, then ignore it and fix the grand scheme.

How Has This Been Tested?
---
Built and provided to the mobile team. Confirmed no build errors.

What process can a PR reviewer use to test or verify this change?
---
Skim the code, but honestly don't look too close it's about to churn.

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify