feat: failure of min difficulty should not add block to list of bad blocks (#5805)
Description
---
Blocks that fail min difficulty should not be added to the list of bad
blocks.

Motivation and Context
---
By adding the blocks to the list of bad blocks we open up the local node
to be spammed to write bad block hashes into the database. These waste
space and they require a write lock on the database. Checking the
difficulty is a very quick and inexpensive check.