fix: don't ban a peer for sending a banned peer (#5843)
Description
---
Removes a check that would ban a peer if they sent a peer that we have
banned.

Motivation and Context
---
The check would ban the source peer during peer sync if they sent a peer
to us that we have banned. This is IMO exploitable, since a peer may not
have connected to the banned peer in question and thus not know if it
should be banned or not.

You could also get a peer banned in this manner:
1. As an evil peer, I connect to a target node A. 
2. I then connect to a number of other nodes and get myself banned. 
3. The nodes I connect to then will ban node A when they send my node
address during sync

How Has This Been Tested?
---
CI

What process can a PR reviewer use to test or verify this change?
---
Not sure it's quite a mission to replicate this.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->