fix: ban peer if it sends bad liveness data (#5844)
Description
---
Bans the peer if they sent invalid liveness data.

Motivation and Context
---
I think there was a change to the chainmetadata struct recently, and
this resulted in a bunch of warnings on my igor node. This PR now bans
the node if they sent data that doesn't deserialize.

How Has This Been Tested?
---
Tested with igor trying to connect to old nodes

What process can a PR reviewer use to test or verify this change?
---
Try to connect to an igor node running 0.52.0-pre.1

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify

<!-- Does this include a breaking change? If so, include this line as a
footer -->
<!-- BREAKING CHANGE: Description what the user should do, e.g. delete a
database, resync the chain -->