refactor(chatffi): remove c rep types (#5829)
Description
---
- Remove the c rep types so clients don't have to deal with unsafe
mutable pointers. It keeps the types used within chat consistent and
creates far less duplication of the same data in different wrappers.
- Update documentation
- Update function and type names
- Try and use better aligned int types. In Rust if we want a u8 we
should request a c_uchar. i64 should be a u_longlong. Previously most
types were being casted down to c_int or c_uint (i32, u32). This
presented a probability for casting problems later on.

Motivation and Context
---
Removing the creps is to help with overall library consistency. The rest
of the changes are improvements for ease of use.

How Has This Been Tested?
---
- CI
- Cucumber
- New unit tests added

What process can a PR reviewer use to test or verify this change?
---
See if the changes make sense. Does the library architecture make sense
to you, could you use it?

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify