refactor(proto)!: remove proto timestamp wrapper types (#5833)
Description
---
We were making use of 3 proto wrapper types. These types deserialize
into option types and in some cases are not what we want.

This PR will update 2 of these types, while saving the third for an
additional PR.

Motivation and Context
---
Use standard rust types, instead of library specific types.

Closes: #5797

How Has This Been Tested?
---
CI

What process can a PR reviewer use to test or verify this change?
---
Pay close attention to changes around timestamps. Could get tricky.
Notably, we were using `Some(0)` in some tests for timestamps. This was
a valid timestamp for us, because it wasn't `None`. Now we're treating
`0` as an invalid timestamp. So any test with `None` became `0` but
tests with `Some(0)` are presumably tests that _should_ have had valid
timestamps but this was the easiest minimal value that was accepted.

<!-- Checklist -->
<!-- 1. Is the title of your PR in the form that would make nice release
notes? The title, excluding the conventional commit
tag, will be included exactly as is in the CHANGELOG, so please think
about it carefully. -->


Breaking Changes
---

- [ ] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [x] Other - Please specify

Messages between new and old nodes are likely to break.