feat(chatffi): get conversationalists (#5849)
Description
---
Add a feature to fetch a list of TariAddresses that a client currently
has conversations with. This is any address a message has been sent to
or received from.

Motivation and Context
---
Without being about to fetch a list of addresses a client has
conversations with, the client can't actually fetch any independent
conversations, as conversations are fetched with an address id.

How Has This Been Tested?
---
New test additions

What process can a PR reviewer use to test or verify this change?
---
Changes are pretty standard throughout. Handler/Service/DB updates.

Breaking Changes
---

- [x] None
- [ ] Requires data directory on base node to be deleted
- [ ] Requires hard fork
- [ ] Other - Please specify