Add future safety
This check ins't needed due to the logic build up before its use but
I've added it so we don't raise and false positives or convern in the
future, and it protects against future changes. I also made not about
the reasons it differs between its counterpart. And some test cases.